+++ pyengine/__init__.py
@@ -1,7 +1,7 @@
 try:
-    from pyengine.Window import Window
+    from pyengine.Window import Window, WindowCallbacks
     from pyengine.Entity import Entity
-    from pyengine.World import World, WorldCallbacks
+    from pyengine.World import World
     from pyengine.Components.PhysicsComponent import CollisionCauses
     from pyengine.Components.ControlComponent import ControlType, Controls, MouseButton
     from pygame import locals as const
@@ -9,5 +9,5 @@ except ModuleNotFoundError:
     pass
 
 __all__ = ["Window", "Entity", "ControlType", "const", "MouseButton", "CollisionCauses",
-           "WorldCallbacks", "World", "Controls"]
 __version__ = "1.2.0"
+           "WindowCallbacks", "World", "Controls"]